synaptics S3320: add support for the proximity state
When the Android Pocket Mode is enabled, the fingerprint reader is
disabled. We'd like to do the same with the touchscreen when double tap
to wake (for example) is enabled. To do this, create the /proc/touchpanel/proximity_state
node: when '1' is written there, the driver will not report anymore
events related to double tap to wake.

Signed-off-by: Roberto Sartori <roberto.sartori.android@gmail.com>